fix(lsp): ansi strip on hover text
