+++ crates/nu-lsp/Cargo.toml
@@ -12,6 +12,7 @@ nu-cli = { path = "../nu-cli", version = "0.102.1" }
 nu-glob = { path = "../nu-glob", version = "0.102.1" }
 nu-parser = { path = "../nu-parser", version = "0.102.1" }
 nu-protocol = { path = "../nu-protocol", version = "0.102.1" }
+nu-utils = { path = "../nu-utils", version = "0.102.1" }
 
 crossbeam-channel = { workspace = true }
 lsp-server = { workspace = true }
